+++ src/components/Page/PageComplete/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {Card, CardText} from 'material-ui/Card';
 import {Markdown} from '../../index';
-import {cyan500} from 'material-ui/styles/colors';
+import {cyan500, grey100} from 'material-ui/styles/colors';
 
 const styles = {
   backgroundColor: cyan500,
@@ -10,7 +10,12 @@ const styles = {
   padding: '10px 15px 30px',
   right: '0px',
   margin: '0',
-  width: '400px',
+  width: '400px'
+};
+
+const textStyles = {
+  color: grey100,
+  fontSize: '1.1em'
 };
 
 export const PageComplete: React.StatelessComponent<{
@@ -20,7 +25,7 @@ export const PageComplete: React.StatelessComponent<{
     {page.completed && page.onPageComplete
         ? <Card style={styles}>
             <CardText>
-                <Markdown>{page.onPageComplete}</Markdown>
+                <Markdown style={textStyles}>{page.onPageComplete}</Markdown>
             </CardText>
           </Card>
         : null