+++ src/components/Page/Tasks/Task.tsx
@@ -1,20 +1,13 @@
 import * as React from 'react';
 import {Markdown} from '../../index';
-// import {TaskCheckbox} from './TaskCheckbox';
+import {TaskCheckbox} from './TaskCheckbox';
 import {ListItem} from 'material-ui/List';
 import {lightGreen200, orange200} from 'material-ui/styles/colors';
 
 function getStatus(
   index: number, taskPosition: number, testRun: boolean
 ): string {
-  switch (true) {
-    case index < taskPosition:
-      return lightGreen200;
-    case testRun:
-      return orange200;
-    default:
-      return 'inherit';
-  }
+  return index < taskPosition ? lightGreen200 : 'inherit';
 }
 
 export const Task: React.StatelessComponent<{
@@ -27,6 +20,7 @@ export const Task: React.StatelessComponent<{
       className='cr-task'
       style={{backgroundColor}}
     >
+      {testRun ? <TaskCheckbox /> : null}
       <span className='cr-task-index'>{index + 1}.</span>
       <div className='cr-task-description'>
         <Markdown >{task.description}</Markdown>